move web3/ocean detection out of App.jsx
