+++ client/src/routes/Faucet.tsx
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react'
 import Route from '../components/templates/Route'
 import Button from '../components/atoms/Button'
 import Spinner from '../components/atoms/Spinner'
-import { User } from '../context/User'
+import { User } from '../context'
 import Web3message from '../components/organisms/Web3message'
 import styles from './Faucet.module.scss'
