+++ client/src/routes/History.tsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import Route from '../components/templates/Route'
 import AssetsUser from '../components/organisms/AssetsUser'
 import Web3message from '../components/organisms/Web3message'
-import { User } from '../context/User'
+import { User } from '../context'
 
 export default class History extends Component {
     public render() {
