+++ client/src/routes/Search.tsx
@@ -3,7 +3,7 @@ import queryString from 'query-string'
 import { Logger } from '@oceanprotocol/squid'
 import Spinner from '../components/atoms/Spinner'
 import Route from '../components/templates/Route'
-import { User } from '../context/User'
+import { User } from '../context'
 import Asset from '../components/molecules/Asset'
 import Pagination from '../components/molecules/Pagination'
 import styles from './Search.module.scss'
